requirements.txt
jupyter notebook
python-dotenv
requests
grok